+++ test/UniswapV3SwapExamples.sol
@@ -1,40 +1,36 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.4.0 <0.9.0;
+
 pragma experimental ABIEncoderV2;
 
 contract UniswapV3SwapExamples {
-    ISwapRouter constant router =
-        ISwapRouter(0xE592427A0AEce92De3Edee1F18E0157C05861564);
+    ISwapRouter constant router = ISwapRouter(0xE592427A0AEce92De3Edee1F18E0157C05861564);
 
-    function swapExactInputSingleHop(
-        address tokenIn,
-        address tokenOut,
-        uint24 poolFee,
-        uint amountIn
-    ) external returns (uint amountOut) {
+    function swapExactInputSingleHop(address tokenIn, address tokenOut, uint24 poolFee, uint256 amountIn)
+        external
+        returns (uint256 amountOut)
+    {
         IERC20(tokenIn).transferFrom(msg.sender, address(this), amountIn);
         IERC20(tokenIn).approve(address(router), amountIn);
 
-        ISwapRouter.ExactInputSingleParams memory params = ISwapRouter
-            .ExactInputSingleParams({
-                tokenIn: tokenIn,
-                tokenOut: tokenOut,
-                fee: poolFee,
-                recipient: msg.sender,
-                deadline: block.timestamp,
-                amountIn: amountIn,
-                amountOutMinimum: 0,
-                sqrtPriceLimitX96: 0
-            });
+        ISwapRouter.ExactInputSingleParams memory params = ISwapRouter.ExactInputSingleParams({
+            tokenIn: tokenIn,
+            tokenOut: tokenOut,
+            fee: poolFee,
+            recipient: msg.sender,
+            deadline: block.timestamp,
+            amountIn: amountIn,
+            amountOutMinimum: 0,
+            sqrtPriceLimitX96: 0
+        });
 
         amountOut = router.exactInputSingle(params);
     }
 
-    function swapExactInputMultiHop(
-        bytes calldata path,
-        address tokenIn,
-        uint amountIn
-    ) external returns (uint amountOut) {
+    function swapExactInputMultiHop(bytes calldata path, address tokenIn, uint256 amountIn)
+        external
+        returns (uint256 amountOut)
+    {
         IERC20(tokenIn).transferFrom(msg.sender, address(this), amountIn);
         IERC20(tokenIn).approve(address(router), amountIn);
 
@@ -55,60 +51,50 @@ interface ISwapRouter {
         address tokenOut;
         uint24 fee;
         address recipient;
-        uint deadline;
-        uint amountIn;
-        uint amountOutMinimum;
+        uint256 deadline;
+        uint256 amountIn;
+        uint256 amountOutMinimum;
         uint160 sqrtPriceLimitX96;
     }
 
     /// @notice Swaps amountIn of one token for as much as possible of another token
     /// @param params The parameters necessary for the swap, encoded as ExactInputSingleParams in calldata
     /// @return amountOut The amount of the received token
-    function exactInputSingle(ExactInputSingleParams calldata params)
-        external
-        payable
-        returns (uint amountOut);
+    function exactInputSingle(ExactInputSingleParams calldata params) external payable returns (uint256 amountOut);
 
     struct ExactInputParams {
         bytes path;
         address recipient;
-        uint deadline;
-        uint amountIn;
-        uint amountOutMinimum;
+        uint256 deadline;
+        uint256 amountIn;
+        uint256 amountOutMinimum;
     }
 
     /// @notice Swaps amountIn of one token for as much as possible of another along the specified path
     /// @param params The parameters necessary for the multi-hop swap, encoded as ExactInputParams in calldata
     /// @return amountOut The amount of the received token
-    function exactInput(ExactInputParams calldata params)
-        external
-        payable
-        returns (uint amountOut);
+    function exactInput(ExactInputParams calldata params) external payable returns (uint256 amountOut);
 }
 
 interface IERC20 {
-    function totalSupply() external view returns (uint);
+    function totalSupply() external view returns (uint256);
 
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 
-    function transfer(address recipient, uint amount) external returns (bool);
+    function transfer(address recipient, uint256 amount) external returns (bool);
 
-    function allowance(address owner, address spender) external view returns (uint);
+    function allowance(address owner, address spender) external view returns (uint256);
 
-    function approve(address spender, uint amount) external returns (bool);
+    function approve(address spender, uint256 amount) external returns (bool);
 
-    function transferFrom(
-        address sender,
-        address recipient,
-        uint amount
-    ) external returns (bool);
+    function transferFrom(address sender, address recipient, uint256 amount) external returns (bool);
 
-    event Transfer(address indexed from, address indexed to, uint value);
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
 }
 
 interface IWETH is IERC20 {
     function deposit() external payable;
 
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 }
